refactor(gui): split the e2e fixture out of the smoke test
The smoke test crossed the 600-line cap once the command palette flow landed
on top of the locale-sourced assertions. The gate asks for a split by
responsibility rather than shaved lines, and there is a clean seam: the
hermetic ledger fixture, the locale readers and the Electron teardown build
the world, while the test asserts what the app does with it.
